# swagger.yaml
openapi: 3.0.3
info:
  title: rade API
  description: |-
    보스 몬스터를 잡은 결과를 Ranking으로 보여주는 서비스입니다
servers:
  - url: http://localhost:3000
tags:
  - name: user
    description: 유저에 대한 CR API
    externalDocs:
      description: 유저를 생성하고 조회합니다
      url: http://localhost:3000/user

  - name: bossRaid
    description: boss Raid 대한 API 입니다
    externalDocs:
      description: boss Raid에 관한 데이터를 제공합니다
      url: http://localhost:3000/bossRaid
paths:
  /user:
    post:
      tags:
        - user
      summary: 유저 등록
      description: 새로운 유저를 등록합니다
      operationId: createUser
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    example: 1
        '500':
          description: Internal Server Error
  /user/{userId}:
    get:
      tags:
        - user
      summary: 유저의 보스 레이드 정보 조회
      description: 유저의 레이드 정보를 가져옵니다
      operationId: getUserRaidRecords
      parameters:
        - name: userId
          in: path
          description: 조회하는 레코드의 사용자 아이디
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalScore:
                    type: integer
                    example: 30
                  raidRecords:
                    type: array
                    items:
                      type: object
                      properties:
                        raidRecordId:
                          type: integer
                          example: 1
                        score:
                          type: integer
                          example: 30
                        enterTime:
                          type: string
                          format: date-time
                          example: '2022-09-16 18:00'
                        endTime:
                          type: string
                          format: date-time
                          example: '2022-09-16 18:03'
        '500':
          description: Internal Server Error
  /bossRaid:
    get:
      tags:
        - bossRaid
      summary: 보스 raid 정보를 가져옵니다
      description: 보스 레이드 시도가 가능한지 알 수 있습니다
      operationId: getBossState
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
              examples:
                result1:
                  value:
                    canEnter: true
                result2:
                  value:
                    canEnter: false
                    userId: 1
        '500':
          description: Internal Server Error
  /bossRaid/topRankerList:
    get:
      tags:
        - bossRaid
      summary: 보스 raid ranking 정보
      description: 보스 raid에서 나의 ranking과 전체 랭킹을 조회합니다
      operationId: getRankings
      requestBody:
        description: raid 랭킹 조회시 필요합니다
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ranking'
        '500':
          description: Internal Server Error
  /bossRaid/enter:
    post:
      tags:
        - bossRaid
      summary: 보스 raid를 시작합니다.
      description: 보스 레이드를 시작합니다
      operationId: enterBossRaid
      requestBody:
        description: 보스 레이드 시작시 필요합니다
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
                level:
                  type: integer
                  example: 1
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  isEntered:
                    type: boolean
                    example: true
                  raidRecordId:
                    type: integer
                    example: 1
        '500':
          description: Internal Server Error
  /bossRaid/end:
    patch:
      tags:
        - bossRaid
      summary: 보스 raid를 종료합니다.
      description: 보스 레이드를 종료합니다
      operationId: finishBossRaid
      requestBody:
        description: 보스 레이드 종료시 필요합니다
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
                raidRecordId:
                  type: integer
                  example: 1
        required: true
      responses:
        '200':
          description: Successful operation
        '500':
          description: Internal Server Error

components:
  schemas:
    ranking:
      type: object
      properties:
        topRankerInfoList:
          type: array
          items:
            type: object
            properties:
              ranking:
                type: integer
                example: 0
              userId:
                type: integer
                example: 1
              totalScore:
                type: integer
                example: 130
        myRankingInfo:
          type: object
          properties:
            ranking:
              type: integer
              example: 0
            userId:
              type: integer
              example: 1
            totalScore:
              type: integer
              example: 130